Small semaphore fix - work in progress

+++ h/sem.hpp
@@ -28,7 +28,7 @@ private:
 
     unsigned init;
 
-    unsigned val;
+    long int val;
 };
